+++ frontend/app.json
@@ -9,7 +9,7 @@
     "splash": {
       "image": "./assets/splash.png",
       "resizeMode": "contain",
-      "backgroundColor": "#ffffff"
+      "backgroundColor": "#64A759"
     },
     "ios": {
       "supportsTablet": true
